Modify

+++ Proxy/Task3/ScoreStorage.sol
@@ -4,13 +4,13 @@ pragma solidity >= 0.7.0 < 0.9.0;
 
 contract ScoreStorage {
 
-  mapping(bytes32=>uint) uints;
+  mapping(bytes32=>uint256) uints;
   
-  function setUints(bytes32 key, uint score) public {
+  function setUints(bytes32 key, uint256 score) public {
     uints[key] = score;
   }
   
-  function getUints(bytes32 key) public view returns(uint) {
+  function getUints(bytes32 key) public view returns(uint256) {
     return uints[key];
   }
 }
